move test_api_client to tests/services
