+++ tests/views/test_score_encoding.py
@@ -33,7 +33,7 @@ from django.urls import reverse
 
 from base.tests.factories.user import UserFactory
 from internship.models.score_encoding_utils import APDS, MIN_APDS, MAX_APDS
-from internship.tests.views.test_api_client import MockAPI
+from internship.tests.services.test_api_client import MockAPI
 
 
 @override_settings(URL_INTERNSHIP_API='url_test_api')
@@ -116,14 +116,14 @@ class TestScoreEncoding(TestCase):
         self.assertEqual(response.status_code, 302)
         self.assertEqual(messages_items[0].level, SUCCESS)
 
-    @mock.patch('internship.tests.views.test_api_client.MockAPI.scores_affectation_uuid_validate_post')
+    @mock.patch('internship.tests.services.test_api_client.MockAPI.scores_affectation_uuid_validate_post')
     def test_score_validation_success(self, mock_validation_response):
         mock_validation_response.return_value = {}, 204, {}
         url = reverse('internship_score_encoding_validate', kwargs={'affectation_uuid': str(uuid.uuid4())})
         json_response = self.client.get(url).json()
         self.assertDictEqual(json_response, {})
 
-    @mock.patch('internship.tests.views.test_api_client.MockAPI.scores_affectation_uuid_validate_post')
+    @mock.patch('internship.tests.services.test_api_client.MockAPI.scores_affectation_uuid_validate_post')
     def test_score_validation_fail(self, mock_validation_response):
         mock_validation_response.return_value = {'error': 'error'}, 404, {}
         url = reverse('internship_score_encoding_validate', kwargs={'affectation_uuid': str(uuid.uuid4())})
